+++ packages/block-editor/src/store/selectors.js
@@ -23,6 +23,7 @@ import {
 	getParsedPattern,
 	getGrammar,
 	mapUserPattern,
+	isPatternOverride,
 } from './utils';
 import { orderBy } from '../utils/sorting';
 import { STORE_NAME } from './constants';
@@ -2481,6 +2482,7 @@ export const getInserterItems = createRegistrySelector( ( select ) =>
 			)
 				? unlock( select( STORE_NAME ) )
 						.getReusableBlocks()
+						.filter( ( block ) => ! isPatternOverride( block ) )
 						.map( buildReusableBlockInserterItem )
 				: [];