+++ packages/patterns/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### New Features
+
+-   Add the private `createPatternOverride` action and `getPatternOverride` selector to create and find the editable copy of a registered pattern ([#YYYYY](https://github.com/WordPress/gutenberg/pull/YYYYY)).
+
 ### Internal
 
 -   Use the `.jsx` extension for JavaScript source files that contain JSX ([#80990](https://github.com/WordPress/gutenberg/pull/80990)).